moved v2 routes into v2 directory so it can work alongside v1 routes with FastAPI
