+++ app/router/v2/locations.py
@@ -1,9 +1,9 @@
 from fastapi import HTTPException, Request
 
-from ..enums.sources import Sources
-from ..models.location import LocationResponse as Location
-from ..models.location import LocationsResponse as Locations
-from . import router
+from ...enums.sources import Sources
+from ...models.location import LocationResponse as Location
+from ...models.location import LocationsResponse as Locations
+from .. import router
 
 
 @router.get("/locations", response_model=Locations, response_model_exclude_unset=True)
