+++ app/router/v2/sources.py
@@ -1,5 +1,5 @@
-from ..data import data_sources
-from . import router
+from ...data import data_sources
+from .. import router
 
 
 @router.get("/sources")
